feat: Retain a backlog of log data when no follower is attached (#1)
* feat: Retain a backlog of log data when no follower is attached.

When `gopmctl logs` connects to a process, a new logger is created and
attached to the existing composite logger that follows the process's
`stdout` or `stderr`. Unless you have configured the process to also
write logs to a file you won't be able to see content emitted when you
aren't observing via `gopmctl logs`.

This introduces a ring-buffer structure, that captures a fixed amount of
history from a process at all times. When `gopmctl logs` is attached, it
will read the contents of the ring-buffer and write those to the gRPC
response before attaching the new logger that follows logs from that
point on.

When content is read from the ring-buffer, we seek to the first newline
(if we can find one) and use the remaining content. While this does toss
some bytes on the floor, it keeps the output looking nice and keeps the
write path simple.

By default, I've established a fixed 10KB of data for each process. This
should be more than enough for most situations. We can consider making
this configurable in the future.

This also teaches `gopmctl logs` to act like `tail -n` by introducing an
`-n` flag. This instructs the supervisor to return the last N lines of
the backlog when `logs` is called. From that point on `logs` will act as
a normal follower.

* fix: Forgot to rename the constructor for RingBuffer.

* fix: Tie the backlogs into the log clearing behavior.

* fix: Ensure stderr redirection works with the backlog.