feat: Persistent logger for process.
Loggers currently get thrown away every time the Process restarts. This
causes issues for `gopmctl logs` following, because when the Process
restarts, it remains attached to a logger that won't be receiving logs
from the new Process command instance.

This establishes a persistent `logger.CompositeLogger` that remains for
the entire lifecycle of a Process. Each time a Process's command is
started, new loggers are created and mounted to the persistent logger.

You are now able to leave `gopmctl logs <process>` running and do things
like `gopmctl stop <process>` and `gopmctl start <process>` and still
see log output.