feat(config): Top-level environment variables (#14)
* feat(config): Top-level environment variables.

This introduces a top-level environment variable definition. It looks
like this:

    environment:
        TELEGRAF_BIN: telegraf

    programs:
        ...

The original definition only had space for `path` sub-key, but this can
be generalized to variable definitions as long as we perform variable
expansion on the values. This environment is applied before overlaying
the environment defined by each process.

Since we're now doing variable expansion in these top-level environment
variables, I've also added the expansion to the process environments.

The config types for process environments have been switched to use
`env.KeyValue` instead of `map[string]string` to preserve ordered map
behavior of YAML. This will prevent weird surprises in the precedence
order in how environment expansion is applied to these values.

The github.com/goccy/go-yaml dependency was upgraded to v1.7.17 to
correct an issue with ordered-map unmarshaling.

* fix(process): Apply .YML environment before .ENV and OS environment.