fix: improve backlog behavior
Currently, if you ask for lines of backlog when the backlog isn't full,
it returns a bunch of zero bytes because the ring buffer doesn't
maintain an idea of how many bytes are in it.

This change changes the ring buffer code so that it grows the buffer
on demand and keeps track of the overall file offset, which might be
useful in the future.

We also change the code to maintain 1MiB of backlog rather than the
rather-too-small 10KiB, and provide an option to avoid the blocking
tail behaviour, making it easier to run tools such as grep on the backlog
data.

Also remove the `ChanLogger`  implementation in favour of a somewhat
simpler approach using `io.Pipe` that doesn't require the somewhat abstraction-breaking
`PutBuffer` method.