move towards end-to-end testing
We'd like to be able to have tests that give assurance that `gopm` works
as advertised, so this is a start in that direction. It uses [testscript](https://pkg.go.dev/github.com/rogpeppe/go-internal/testscript)
to run end-to-end tests without significantly disrupting the existing
logic in the main packages.

Also:
- remove the "group" functionality that's not implemented currently anyway,
- avoid printing ANSI escape codes when `TERM=dumb` or the output isn't to a terminal.
- avoid using an external diff package when `reflect.DeepEqual` will do just as well.

I have no idea why the `rpc/javascript/package-lock.json` file grew so much when
I ran `make webgui/js/bundle.js` to regenerate the JS when removing the unused
`SignalProcessGroup` call from the gRPC API.