only support a single log file
Currently gopm supports independent stdout and stderr log files,
but this is of marginal use and complicates all the logic.
If a separate log file for one or the other is desired, the
user can always use file redirection.

This PR removes support for independent stdout/stderr log files,
using only a single combined stdout/stderr log file per process.

Also add a test and fix a bug that exposed in the `lastNLines` function.