process: move ForEach operations inside process manager
Moving towards a major refactor of the `process` package, as a
first step, we move all the various operations that iterate
over processes to inside the process manager, meaning
that we can unexport the `Process` type and all its methods.

We also move the `process.Stat` function and its dependencies
out into its own package as it has nothing to do with the rest
of that package's functionality (it's only used by `gopmctl`,
in a somewhat dubious manner that assumes that `gopmctl`
is running on the same machine as the `gopm` it's talking to).

We remove the topological sorting code too, as it doesn't actually
help in doing the job it's supposed to be doing.