move shell to config
Having the shell as a command-line value that defaults
to $SHELL doesn't really make sense, as the interpretation
of the commands in a `gopm` config file shouldn't vary depending
on the environment they're run in. Instead, make the shell
a property of each program, defaulting to `/bin/sh`.

Also remove the restart-when-binary-changed functionality
as it's currently somewhat broken and doesn't really make sense
in general when commands are shell scripts. It can be added
back later with rethought semantics.