better config verification
Verify cron values when parsing; also verify depends_on relationships,
meaning that we can assume later in the code that there are no cyclic
dependencies to get us into recursive trouble.

Also remove the `priority` field from the configuration as it's of marginal
use since we now have `depends_on` which wasn't present in the
original `supervisord` configuration from which this is derived.