process: rewrite process management
We rewrite the process management to use a more
disciplined approach to concurrency. There's a main
goroutine for each process that monitors any process
that's currently active, and another goroutine that's
responsible for coordinating updates to the configuration.

Logging isn't hooked up and restarts aren't tested at all,
but this is probably a good stopping point for a review.