better server management
This updates the logic so that the gopm server can be shut down
and updated from within the HTTP web page.
It also tries to prevent the user from accidentally shutting
off access to their own gopm instance by starting it with
no servers or servers that can't be listened on.

It also fixes the config directory name handling so that
absolute paths work properly.

Also update the build process so that it builds a release
binary by default, so plain `go install` will do the right thing,
and change the targets so that `make` will just install the binaries
by default, and use a more specific `$GOPM_DEV` variable
for selecting a dev build (rather than just `$DEV`).