config: move #RuntimeConfig out of #Config
Move the `runtime` field to the top level of the configuration
so the separation of concerns is clearer between gopm-provided
values (`runtime`) and user-specifiable values (`config`).

This avoids the name clash of having `gopm` being both a top
level field and the name of the package.

We also add a `version` field so that we can use it to disambiguate between
versions in future without worrying that some users might have defined it
themselves.