config: add absPath to files
The `#Paths` definition, although a bit more concise, seems
like it's likely to be a bit harder for users to understand.
Instead, add an `absPath` field to `#File` that's filled out
automatically.

Alter the filesystem test to check that it's actually possible to
refer to file paths from within programs and other files.