improve CI tests
We make the tests run on Darwin as well as Ubuntu (let's leave
Windows for now :) ) and reinstate the `go mod tidy` test in the
hope that it's better now.

The main motivation for this is because the new fswatcher package
tests have Darwin-specific code which is hard to test otherwise.

Also fix a race condition in the `gopm` package that was found
by the race detector and run `go fmt`.

Also move the `tools.go` file into the top level package so that
it's automatically formatted when running `go fmt ./...` which it
wasn't before because there were no non-excluded files in that
package.