feat: implement oneshot attribute
The new `oneshot` attribute on a program marks it so that it will never
be marked as running - it will only go into state Exited or Fatal.
This will be useful for programs that implement one-off tasks as
prerequisites for others, where we don't want programs that depend
on them to be started until the task has completed.

Also fix a bug exposed by the new test where a newly created program
in Stopped state would count as Ready for the purposes of a `gopmctl start` command
because the entry wasn't added to the `stateNotifier.state` map.