show an error if gopmctl start or restart fails
This also gives us an opportunity to experiment with returning
rich errors from gRPC calls, which paves the way for returning
syntax errors that can be formatted nicely on the client side.

Also fix an error in the stop logic that would leave a process in Backoff
state when it should be moved to Stopped.