feat: restart dependencies on update if they were already running
This change means that if a dependency is changed by an update
and it wasn't explicitly stopped, that it will be restarted after its configuration
has been updated.

It also significantly improves the dependency tests so that we test the
previously implemented but untested behaviour that
when a program is changed, all the programs that depend on it will
be stopped too. We also fix the `walkChanged` logic that was causing
this functionality to not work previously.