process: refactor configuration updates
Currently when the configuration is being updated, the
updater first interrogates all the processes to see what
state they're in before updating their configuration
and restarting if they weren't previously stopped.
There's an issue with that, which is that if the update
is aborted, that previous state is lost, so processes
won't restart again.

This makes the updater goroutine responsible for notifying
processes that their dependencies are ready, meaning that
processes don't have to watch their own dependencies directly,
which simplifies process creation somewhat and it adds
a new Paused state so that processes themselves remember
if they were previously started.

This fixes starting and stopping so that it preserves the invariant
that if a process was ever in Starting state, unless it's explicitly
stopped, it will restart automatically whenever any of its dependencies
are restarted.