process: avoid the possibility of some commands hanging indefinitely
Currently, if we're trying to stop or start processes and waiting for them
to become stopped or running, that wait will wait indefinitely if
some concurrent command starts or stops one of the processes while
we're waiting, because the wait waits only for the expected state.

Fix this by always waiting for a steady state (e.g. Running, Stopped, Fatal),
and then checking that the steady state is the one that we're expecting.
We avoid a race condition by adding a reply channel to the request
that allows the process to notify the caller when its state has been
initially updated in response to the request.