config: use map for depends_on, not slice
`[string]: true` is a much more natural representation of a set in CUE, and makes
it easy to add dependencies in an orthogonal manner because `a: true` unifies
with `b: true`, whereas `["a"]` does not unify with `["b"]`.

This is arguably a slight syntactic inconvenience, but I think it's worth doing.

Note: this is a backwardly incompatible change.